Wrapper Function To Check Pedigrees For Cycles

The PedigreeCycleCheck R6 class is used to find cycles in a pedigree. The pedigree can be specified via input file or via a tibble. The result consists either of TRUE or FALSE which tells whether a pedigree contains a loop or not. In addition the animals that are part of the loop can be listed.

check_cycle_pedigree(
  ps_pedig_path = NULL,
  pb_report_cycle = FALSE,
  ps_delim = "|",
  pn_id_col = 1,
  pn_sire_col = 2,
  pn_dam_col = 3,
  pcol_types = NULL,
  ptbl_pedigree = NULL
)

Arguments

ps_pedig_path
path to the pedigree input file

pb_report_cycle
flag, whether animals in cycle are reported

ps_delim
delimiting character of columns in input file

pn_id_col
column index of animals in input file

pn_sire_col
column index of sires in input file

pn_dam_col
column indes of dams in input file

pcol_types
data types of columns in input file, passed to readr

ptbl_pedigree
pedigree given as tbl

Details

Cycles are searched using depth-first-traversals of the pedigree.

Examples

if (FALSE) {
check_cycle_pedigree(ps_pedig_path = system.file('extdata','data_sample2.csv', package = 'qprppedigree'))
}