Create Pedigree Report

The checks in this package are done for a given pedigree given by ps_pedigree_path. The results of the checks are summarised in a report which is generated by this function.

create_pedigree_report(
  ps_pedigree_path,
  ps_report_outdir = ".",
  ps_report_rmd = paste0(format(Sys.time(), "%Y%m%d%H%M%S"),
    "_qpdt_pedigree_report.Rmd", collapse = ""),
  pl_params = NULL
)

Arguments

ps_pedigree_path
path to the pedigree file

ps_report_outdir
report output directory

ps_report_rmd
name of the pedigree report file

pl_params
list of parameters to be included in the report

Examples

if (FALSE) {
s_pedigree_path <- file.path(here::here(), 'inst', 'extdata', 'PopReport_OST_mit_20210318.csv_adaptfin5.csv')
create_pedigree_report(ps_pedigree_path = system.file('extdata',
    'PopReport_SN_ohne_20210115.csv_adaptfin2.csv',package = 'qpdt'),
  ps_report_outdir = file.path(here::here(), 'scratch', 'report_out'))
}